# docs/swagger.yaml
definitions:
  dbmodel.Cat:
    properties:
      age:
        type: integer
      breed:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      visits:
        items:
          $ref: '#/definitions/dbmodel.Visit'
        type: array
      weigth:
        type: integer
    type: object
  dbmodel.Treatment:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      visit:
        $ref: '#/definitions/dbmodel.Visit'
      visit_id:
        type: integer
    type: object
  dbmodel.Visit:
    properties:
      cat:
        $ref: '#/definitions/dbmodel.Cat'
      cat_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      motif:
        type: string
      treatments:
        items:
          $ref: '#/definitions/dbmodel.Treatment'
        type: array
      updated_at:
        type: string
      veterinaire:
        type: string
    type: object
...
  models.VisitRequest:
    properties:
      date:
        type: string
      motif:
        type: string
      veterinaire:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /cats/{id}/visits:
    get:
      parameters:
      - description: Cat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dbmodel.Visit'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get visits by Cat ID
      tags:
      - visits
...
  /visits:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dbmodel.Visit'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all visits
      tags:
      - visits
    post:
      consumes:
      - application/json
      parameters:
      - description: Visit payload
        in: body
        name: visit
        required: true
        schema:
          $ref: '#/definitions/models.VisitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dbmodel.Visit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new visit
      tags:
      - visits
  /visits/{id}:
    delete:
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a visit
      tags:
      - visits
    get:
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dbmodel.Visit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a visit by ID
      tags:
      - visits
    put:
      consumes:
      - application/json
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Visit payload
        in: body
        name: visit
        required: true
        schema:
          $ref: '#/definitions/models.VisitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dbmodel.Visit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a visit
      tags:
      - visits
...
  /visits/filter:
    get:
      parameters:
      - description: Motif
        in: query
        name: motif
        type: string
      - description: Veterinaire
        in: query
        name: veterinaire
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dbmodel.Visit'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Filter visits by motif or veterinaire
      tags:
      - visits
swagger: "2.0"